Crate drone_cortexm

ARM® Cortex®-M platform crate for Drone, an Embedded Operating System.

Supported Cores

Architecture	Core name	Rust target	cortexm_core config flag

ARMv7-M	ARM® Cortex®-M3 r0p0	thumbv7m-none-eabi	cortexm3_r0p0

ARMv7-M	ARM® Cortex®-M3 r1p0	thumbv7m-none-eabi	cortexm3_r1p0

ARMv7-M	ARM® Cortex®-M3 r1p1	thumbv7m-none-eabi	cortexm3_r1p1

ARMv7-M	ARM® Cortex®-M3 r2p0	thumbv7m-none-eabi	cortexm3_r2p0

ARMv7-M	ARM® Cortex®-M3 r2p1	thumbv7m-none-eabi	cortexm3_r2p1

ARMv7E-M	ARM® Cortex®-M4 r0p0	thumbv7em-none-eabi	cortexm4_r0p0

ARMv7E-M	ARM® Cortex®-M4 r0p1	thumbv7em-none-eabi	cortexm4_r0p1

ARMv7E-M	ARM® Cortex®-M4F r0p0	thumbv7em-none-eabihf	cortexm4f_r0p0

ARMv7E-M	ARM® Cortex®-M4F r0p1	thumbv7em-none-eabihf	cortexm4f_r0p1

ARMv8-M	ARM® Cortex®-M33 r0p2	thumbv8m.main-none-eabi	cortexm33_r0p2

ARMv8-M	ARM® Cortex®-M33 r0p3	thumbv8m.main-none-eabi	cortexm33_r0p3

ARMv8-M	ARM® Cortex®-M33 r0p4	thumbv8m.main-none-eabi	cortexm33_r0p4

ARMv8-M	ARM® Cortex®-M33F r0p2	thumbv8m.main-none-eabihf	cortexm33f_r0p2

ARMv8-M	ARM® Cortex®-M33F r0p3	thumbv8m.main-none-eabihf	cortexm33f_r0p3

ARMv8-M	ARM® Cortex®-M33F r0p4	thumbv8m.main-none-eabihf	cortexm33f_r0p4

Rust target triple and cortexm_core config flag should be set at the application level according to this table.

Documentation

Drone Book

API documentation

Usage

Add the crate to your Cargo.toml dependencies:

[dependencies]
drone-cortexm = { version = "0.14.0", features = [...] }

Add or extend std feature as follows:

[features]
std = ["drone-cortexm/std"]

Modules

drv
Drivers for core ARM Cortex-M peripherals.

fib
The Fibers module.

map
Core ARM Cortex-M register and exception mappings.

proc_loop
This module provides interface to wrap a stackful synchronous code into an asynchronous command loop.

processor
Common utility functions for working with ARM Cortex-M processors.

reg
The Memory-Mapped Registers module.

reg
The Memory-Mapped Registers module.

sv
The Supervisor module.

swo
Single Wire Output interface.

thr
The Threads module.

thr
The Threads module.

Macros

reg