Module drone_cortexm::map::reg::mpu::rasr

Defines the size and access behavior of the region identified by RNR, and enables that region.

Structs

Ap
Access permission.

B
Memory attribute.

C
Memory attribute.

Enable
Region enable bit.

Hold
Defines the size and access behavior of the region identified by RNR, and enables that region.

Reg
Defines the size and access behavior of the region identified by RNR, and enables that region.

S
Shareable memory attribute.

Size
Size of the MPU protection region.

Srd
Subregion disable bits.

Tex
Memory attribute.

Val
Defines the size and access behavior of the region identified by RNR, and enables that region.

Xn
Instruction access disable bit.