Module drone_cortexm::map::reg::mpu::type_

The MPU Type Register indicates how many regions the MPU support. Software can use it to determine if the processor implements an MPU.

Structs

Dregion
Number of regions supported by the MPU. If this field reads-as-zero the processor does not implement an MPU.

Hold
The MPU Type Register indicates how many regions the MPU support. Software can use it to determine if the processor implements an MPU.

Iregion
Instruction region.

Reg
The MPU Type Register indicates how many regions the MPU support. Software can use it to determine if the processor implements an MPU.

Separate
Indicates support for separate instruction and data address maps.

Val
The MPU Type Register indicates how many regions the MPU support. Software can use it to determine if the processor implements an MPU.