Module drone_cortexm::map::reg::scb::bfsr

BusFault Status Register.

Structs

Bfarvalid
BFAR has valid contents.

Hold
BusFault Status Register.

Ibuserr
A bus fault on an instruction prefetch has occurred. The fault is signaled only if the instruction is issued.

Impreciserr
Imprecise data access error has occurred.

Lsperr
A bus fault occurred during FP lazy state preservation.

Preciserr
A precise data access error has occurred, and the processor has written the faulting address to the BFAR.

Reg
BusFault Status Register.

Stkerr
A derived bus fault has occurred on exception entry.

Unstkerr
A derived bus fault has occurred on exception return.

Val
BusFault Status Register.