Struct drone_cortexm::map::reg::scb::dfsr::Hold

pub struct Hold<'a, _T: RegTag> { /* fields omitted */ }

Debug Fault Status Register.

Implementations

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn external(&self) -> bool

Indicates a debug event generated because of the assertion of an external debug request.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_external(&mut self) -> &mut Self

Indicates a debug event generated because of the assertion of an external debug request.

pub fn clear_external(&mut self) -> &mut Self

Indicates a debug event generated because of the assertion of an external debug request.

pub fn toggle_external(&mut self) -> &mut Self

Indicates a debug event generated because of the assertion of an external debug request.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn vcatch(&self) -> bool

Indicates triggering of a Vector catch.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_vcatch(&mut self) -> &mut Self

Indicates triggering of a Vector catch.

pub fn clear_vcatch(&mut self) -> &mut Self

Indicates triggering of a Vector catch.

pub fn toggle_vcatch(&mut self) -> &mut Self

Indicates triggering of a Vector catch.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn dwttrap(&self) -> bool

Indicates a debug event generated by the DWT.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_dwttrap(&mut self) -> &mut Self

Indicates a debug event generated by the DWT.

pub fn clear_dwttrap(&mut self) -> &mut Self

Indicates a debug event generated by the DWT.

pub fn toggle_dwttrap(&mut self) -> &mut Self

Indicates a debug event generated by the DWT.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn bkpt(&self) -> bool

Indicates a debug event generated by BKPT instruction execution or a breakpoint match in FPB.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_bkpt(&mut self) -> &mut Self

Indicates a debug event generated by BKPT instruction execution or a breakpoint match in FPB.

pub fn clear_bkpt(&mut self) -> &mut Self

Indicates a debug event generated by BKPT instruction execution or a breakpoint match in FPB.

pub fn toggle_bkpt(&mut self) -> &mut Self

Indicates a debug event generated by BKPT instruction execution or a breakpoint match in FPB.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn halted(&self) -> bool

Indicates a debug event generated by either.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_halted(&mut self) -> &mut Self

Indicates a debug event generated by either.

pub fn clear_halted(&mut self) -> &mut Self

Indicates a debug event generated by either.

pub fn toggle_halted(&mut self) -> &mut Self

Indicates a debug event generated by either.