Module drone_cortexm::map::reg::scb::hfsr

HardFault Status Register.

Structs

Debugevt
Debug event has occurred. The Debug Fault Status Register has been updated.

Forced
Indicates that a fault with configurable priority has been escalated to a HardFault exception, because it could not be made active, because of priority or because it was disabled.

Hold
HardFault Status Register.

Reg
HardFault Status Register.

Val
HardFault Status Register.

Vecttbl
Indicates when a fault has occurred because of a vector table read error on exception processing.