Struct drone_cortexm::map::reg::scb::mmfsr::Reg

pub struct Reg<_T: RegTag> {
    pub mmarvalid: Mmarvalid<_T>,
    pub mlsperr: Mlsperr<_T>,
    pub mstkerr: Mstkerr<_T>,
    pub munstkerr: Munstkerr<_T>,
    pub daccviol: Daccviol<_T>,
    pub iaccviol: Iaccviol<_T>,
}

MemManage Status Register.

Fields
mmarvalid: Mmarvalid<_T>

MMFAR has valid contents.

mlsperr: Mlsperr<_T>

A MemManage fault occurred during FP lazy state preservation.

mstkerr: Mstkerr<_T>

A derived MemManage fault occurred on exception entry.

munstkerr: Munstkerr<_T>

A derived MemManage fault occurred on exception return.

daccviol: Daccviol<_T>

Data access violation. The MMFAR shows the data address that the load or store tried to access.

iaccviol: Iaccviol<_T>

MPU or Execute Never (XN) default memory map access violation on an instruction fetch has occurred. The fault is signalled only if the instruction is issued.