Struct drone_cortexm::map::reg::scb::Bfsr

pub struct Bfsr<_T: RegTag> {
    pub bfarvalid: Bfarvalid<_T>,
    pub lsperr: Lsperr<_T>,
    pub stkerr: Stkerr<_T>,
    pub unstkerr: Unstkerr<_T>,
    pub impreciserr: Impreciserr<_T>,
    pub preciserr: Preciserr<_T>,
    pub ibuserr: Ibuserr<_T>,
}

BusFault Status Register.

Fields
bfarvalid: Bfarvalid<_T>

BFAR has valid contents.

lsperr: Lsperr<_T>

A bus fault occurred during FP lazy state preservation.

stkerr: Stkerr<_T>

A derived bus fault has occurred on exception entry.

unstkerr: Unstkerr<_T>

A derived bus fault has occurred on exception return.

impreciserr: Impreciserr<_T>

Imprecise data access error has occurred.

preciserr: Preciserr<_T>

A precise data access error has occurred, and the processor has written the faulting address to the BFAR.

ibuserr: Ibuserr<_T>

A bus fault on an instruction prefetch has occurred. The fault is signaled only if the instruction is issued.