Module drone_cortexm::map::reg::scb::ufsr

UsageFault Status Register.

Structs

Divbyzero
Divide by zero error has occurred.

Hold
UsageFault Status Register.

Invpc
An integrity check error has occurred on EXC_RETURN.

Invstate
Instruction executed with invalid EPSR.T or EPSR.IT field.

Nocp
A coprocessor access error has occurred. This shows that the coprocessor is disabled or not present.

Reg
UsageFault Status Register.

Unaligned
Unaligned access error has occurred.

Undefinstr
The processor has attempted to execute an undefined instruction.

Val
UsageFault Status Register.