Module drone_cortexm::reg::field

Memory-mapped register fields module.

See the top-level module documentation for details.

Re-exports

pub use drone_core::reg::field::*;

Traits

RRRegFieldBitBand
Readable single-bit field of readable register located in the peripheral bit-band region.

WRwRegFieldAtomic
Atomic operations for writable field of read-write register.

WRwRegFieldBitAtomic
Atomic operations for writable single-bit field of read-write register.

WRwRegFieldBitsAtomic
Atomic operations for writable multiple-bit field of read-write register.

WWRegFieldBitBand
Writable single-bit field of writable register located in the peripheral bit-band region.