Trait drone_cortexm::reg::field::RegField

pub trait RegField<T>: Token + Sync where
    T: RegTag, {
    type Reg: Reg<T>;
    type URegField: RegField<Urt>;
    type SRegField: RegField<Srt>;
    type CRegField: RegField<Crt>;

    pub const OFFSET: usize;
    pub const WIDTH: usize;
    pub fn into_unsync(self) -> Self
    where
        Self: RegField<Urt>,
    { ... }

    pub fn into_sync(self) -> Self
    where
        Self: RegField<Srt>,
    { ... }

    pub fn into_copy(self) -> Self::CRegField
    where
        T: RegAtomic,
    { ... }

    pub fn as_sync(&self) -> &Self::SRegField
    where
        T: RegAtomic,
    { ... }
}

The base trait for a field token of a memory-mapped register.

Associated Types

type Reg: Reg<T>

Parent register token.

type URegField: RegField<Urt>

Corresponding unsynchronized register field token.

type SRegField: RegField<Srt>

Corresponding synchronized register field token.

type CRegField: RegField<Crt>

Corresponding copyable register field token.

Loading content...
Associated Constants

pub const OFFSET: usize

The offset of the field inside the parent register.

pub const WIDTH: usize

The bit-width of the field.

Loading content...
Provided methods

pub fn into_unsync(self) -> Self where
Self: RegField<Urt>,

Converts into unsynchronized register field token.

pub fn into_sync(self) -> Self where
Self: RegField<Srt>,

Converts into synchronized register field token.

pub fn into_copy(self) -> Self::CRegField where
T: RegAtomic,

Converts into copyable register field token.

pub fn as_sync(&self) -> &Self::SRegField where
T: RegAtomic,

Returns a reference to the synchronized field token.

Loading content...