Trait drone_cortexm::reg::field::WoWoRegField

pub trait WoWoRegField<T>: WoWRegField<T> where
    T: RegTag,
    Self::Reg: WoReg<T>, {
    pub fn default_val(&self) -> <Self::Reg as Reg<T>>::Val;

    pub fn store_val(&self, val: <Self::Reg as Reg<T>>::Val);

    pub fn store<F>(&self, f: F)
    where
        F: Fn(&mut <Self::Reg as Reg<T>>::Val);
}

Write-only field of write-only register.

Required methods

pub fn default_val(&self) -> <Self::Reg as Reg<T>>::Val

Creates a new opaque register value, and initializes it with the reset value.

pub fn store_val(&self, val: <Self::Reg as Reg<T>>::Val)

Writes an opaque value val into the register memory.

See also store.

pub fn store<F>(&self, f: F) where
F: Fn(&mut <Self::Reg as Reg<T>>::Val),

Passes the opaque reset value to the closure f, then writes the result of the closure into the register memory.

See also store_val.

Loading content...