Module drone_cortexm::reg::prelude

The Memory-Mapped Registers prelude.

The purpose of this module is to alleviate imports of many common reg traits by adding a glob import to the top of reg heavy modules:

use drone_core::reg::prelude::*;

Structs

Crt
The copyable register token tag.

Srt
The synchronized register token tag.

Urt
The unsynchronized register token tag.

Traits

RRRegField
Readable field of readable register.

RReg
Readable register.

Reg
The base trait for a memory-mapped register token.

RegAtomic
An atomic register token tag.

RegField
The base trait for a field token of a memory-mapped register.

RegHold
Exposed storage for register values.

RegOwned
An owned register token tag.

RegTag
A register token tag.

RoRRegField
Read-only field of readable register.

RoReg
Read-only register.

WReg
Writable register.

WWRegField
Writable field of writable register.

WoReg
Write-only register.

WoWRegField
Write-only field of writable register.

_
Readable single-bit field of readable register.

_
Readable multiple-bit field of readable register.

_
Single-bit register field.

_
Multiple-bits register field.

_
Writable single-bit field of writable register.

_
Writable multiple-bit field of writable register.

_
Write-only field of write-only register.

_
Write-only single-bit field of write-only register.

_
Write-only multiple-bit field of write-only register.

_
Connects Reg with RegHold.

_
Non-atomic operations for read-write register.

_
Atomic operations for writable register.

_
Non-atomic operations for writable register.