Trait drone_cortexm::reg::prelude::_

pub trait _<T>: RegFieldBit<T> + RRRegField<T> where
    T: RegTag,
    Self::Reg: RReg<T>, {
    fn read(&self, val: &<Self::Reg as Reg<T>>::Val) -> bool;

    fn read_bit(&self) -> bool;
}

Readable single-bit field of readable register.

Required methods

fn read(&self, val: &<Self::Reg as Reg<T>>::Val) -> bool

Returns true if the bit is set in val.

fn read_bit(&self) -> bool

Reads the value from the register memory and returns true if the bit is set.

Loading content...