Trait drone_cortexm::reg::RReg

pub trait RReg<T>: Reg<T> where
    T: RegTag, {
    fn load<'a>(&'a self) -> Self::Hold
    where
        Self: RegRef<'a, T>,
    { ... }

    fn load_val(&self) -> Self::Val { ... }

    fn load_bits(&self) -> <Self::Val as Bitfield>::Bits { ... }

    fn as_ptr(&self) -> *const <Self::Val as Bitfield>::Bits { ... }
}

Readable register.

Provided methods

fn load<'a>(&'a self) -> Self::Hold where
Self: RegRef<'a, T>,

Reads the value from the register memory to the exposed value type.

See also load_val, load_bits.

fn load_val(&self) -> Self::Val

Reads the value from the register memory to the opaque value type.

See also load, load_bits.

fn load_bits(&self) -> <Self::Val as Bitfield>::Bits

Reads the value from the register memory to the raw value type.

See also load, load_val.

fn as_ptr(&self) -> *const <Self::Val as Bitfield>::Bits

Returns a raw pointer to the register memory.

See also as_mut_ptr.

Loading content...