Module drone_cortexm::swo

Single Wire Output interface.

This module provides interface for ITM (Instrumentation Trace Macrocell) output through SWO (Single Wire Output) pin, and optionally the respective implementation for drone_core::log facade (via set_log! macro).

Structs

Port
ITM stimulus port handle.

Constants

PORTS_COUNT
Number of ports.

Functions

flush
Blocks until all pending packets are transmitted.

is_enabled
Returns true if the debug probe is connected and listening to the ITM output.

is_port_enabled
Returns true if the debug probe is connected and listening to the output of ITM port number port.

sync
Generates an ITM synchronization packet.

update_prescaler
Updates the SWO prescaler register.