Trait drone_cortexm::thr::ExecOutput

pub trait ExecOutput: Sized + Send {
    type Terminate;
    pub fn terminate(self) -> Self::Terminate;
}

A trait for implementing arbitrary output types for futures passed to ThrExec::exec and ThrExec::add_exec.

Associated Types

type Terminate

The return type of ExecOutput::terminate. Should be either () or !.

Loading content...
Required methods

pub fn terminate(self) -> Self::Terminate

A result handler for an executor. The returned value will not be used, so the only useful types are () and !. The handler may choose to panic on an erroneous value.

Loading content...