Trait drone_cortexm::thr::FutureRootExt

pub trait FutureRootExt: Future {
    pub fn root_wait(self) -> Self::Output;
}

An extension trait for Future that provides root_wait method.

Required methods

pub fn root_wait(self) -> Self::Output

Runs a future to completion on the lowest priority thread.

This method will block the caller until the given future has completed.

WARNING This method will block currently preempted threads. It is recommended to use this method only on the lowest priority thread.

Loading content...