Trait drone_cortexm::thr::StreamRootExt

pub trait StreamRootExt<'a>: Stream {
    fn root_wait(self) -> StreamRootWait<'a, Self>
    where
        Self: Sized;
}

An extension trait for Stream that provides root_wait method.

Required methods

fn root_wait(self) -> StreamRootWait<'a, Self> where
Self: Sized,

Turn a stream into a blocking iterator.

When next is called on the resulting StreamRootWait, the caller will be blocked until the next element of the Stream becomes available.

WARNING The resulting StreamRootWait will be blocking preempted threads. It is recommended to use this method only on the lowest priority thread.

Loading content...