Trait drone_cortexm::thr::ThrExec

pub trait ThrExec: ThrToken {
    pub fn wakeup(self);

    pub fn waker(self) -> Waker;

    pub fn exec<F, O>(self, fut: F)
    where
        F: Future<Output = O> + Send + 'static,
        O: ExecOutput,
    { ... }

    pub fn exec_factory<C, F, O>(self, factory: C)
    where
        C: FnOnce() -> F + Send + 'static,
        F: Future<Output = O> + 'static,
        O: ExecOutput,
    { ... }

    pub fn add_exec<F, O>(self, fut: F)
    where
        F: Future<Output = O> + Send + 'static,
        O: ExecOutput,
    { ... }

    pub fn add_exec_factory<C, F, O>(self, factory: C)
    where
        C: FnOnce() -> F + Send + 'static,
        F: Future<Output = O> + 'static,
        O: ExecOutput,
    { ... }
}

Thread executor.

Required methods

pub fn wakeup(self)

Wakes up the thread.

pub fn waker(self) -> Waker

Returns a handle for waking up a thread.

Loading content...
Provided methods

pub fn exec<F, O>(self, fut: F) where
F: Future<Output = O> + Send + 'static,
O: ExecOutput,

Adds an executor for the future fut to the fiber chain and wakes up the thread immediately.

pub fn exec_factory<C, F, O>(self, factory: C) where
C: FnOnce() -> F + Send + 'static,
F: Future<Output = O> + 'static,
O: ExecOutput,

Adds an executor for the future returned by factory to the fiber chain and wakes up the thread immediately.

pub fn add_exec<F, O>(self, fut: F) where
F: Future<Output = O> + Send + 'static,
O: ExecOutput,

Adds an executor for the future fut to the fiber chain.

The future fut will start polling on the next thread wake-up.

pub fn add_exec_factory<C, F, O>(self, factory: C) where
C: FnOnce() -> F + Send + 'static,
F: Future<Output = O> + 'static,
O: ExecOutput,

Adds an executor for the future returned by factory to the fiber chain.

The future fut will start polling on the next thread wake-up.

Loading content...