Module drone_cortexm::map::reg::scb::dfsr

Debug Fault Status Register.

Structs

Bkpt
Indicates a debug event generated by BKPT instruction execution or a breakpoint match in FPB.

Dwttrap
Indicates a debug event generated by the DWT.

External
Indicates a debug event generated because of the assertion of an external debug request.

Halted
Indicates a debug event generated by either.

Hold
Debug Fault Status Register.

Reg
Debug Fault Status Register.

Val
Debug Fault Status Register.

Vcatch
Indicates triggering of a Vector catch.